Profiled the ELM327 Serial port usage & added NVE checks
Turned off UART echo and extra spaces in the ELM327 output, also checked that the ATPPS settings are correct - as Forscan would have set them - if they are not correct, run Forscan to test your ELM327 first (HS & MS CANBUS via the physical switch) and they should then be corrected.  There's a hint there in case you want to do it yourself, but you need to read the ELM327 manual.