Added significant retry logic when writing to the Cluster
If the CANBUS being used has a lot of noise on it, messages may not be received, so the software now checks all write CRC values and attempts to restart a write operation if any communication abnormalities are detected.  Reads will still fail at the first sign of a problem, because that will prevent any write options from being enabled in the software.  I.E. It's best never to start a write operation on a CANBUS that is reporting read errors.