Add independent FGX mode
The FGX has a smart data link connector between the Instrument Cluster and the OBD2 port, and it increases the MS-CAN (125K BUAD) from the Cluster to HS-CAN (500K BAUD) - so this new option can be used to make the ELM327 dongle use the higher speed even when the modified ELM's physical switch is set to the medium speed.  Note though, if the Cluster is out of the car (on the bench) then the FGII mode should still be used, since the SDLC is no longer increasing the BAUD rate.