Backup & later Repair your ICC
[This is an updated version which (only) removes an unneeded shared library.]

Upon inserting the USB stick into the ICC, waiting a bit, and pressing the "Recore" button in the ICCComms.exe program, all factory packages installed on the unit will be copied off the unit and onto the USB stick.  Once this is done, a log file will be generated with a list of all corrupt files.  On a working unit, this list will be empty - and you will be left with a good backup of the software installed.  Non-nav units should take ~20 minutes (should you not have a working unit, and can't see the progress on the screen).  Nav units will take about an hour.  Blinky USB LEDs may stop once the copy is done, but it's still doing the log/corruption-check at that point, so give it ~20 minutes more.  There is no need to reset the ICC when finished, so this process is "safe" to run on a broken unit which has never lost battery power.

Upon inserting the USB stick into the ICC some time later, waiting a bit, and pressing the "Recore" button, it will detect that the backup is already there, and it will check all the files on the unit, any corrupt ones found will have the USB version checked and if it matches the version's manifest checksum on the unit, it will be copied back to the unit to repair the corruption.  The copied file on the unit will then be checked again & copied again if it still has problems - up to 15 times, in an attempt to find a good spot in the flash memory.

I've written the program this way (I.E. check the USB copy version) so it won't put garbage or someone's attempt to "upgrade" certain files, on the unit.  It will only repair using the correct files.  If a repair was made, and it looks successful (none got to 15 retries) it will reboot automatically.