FG MK2 High Series Navi Map Updater
This package backs-up your current maps into a folder it creates on the thumbdrive called "oldcontent" when you run it with no new maps (a freshly formatted FAT32 USB stick, with just the expanded archive on it).

If you have new maps (*.fbl, *.poi or *.spc [btw it doesn't allow filenames with spaces]) in the thumbdrive root, it attempts to patch your navi exe to remove the license check.  It will tell you if it can or can't do it to the version installed.  Then it makes a backup of the maps (as before) and copies new maps over.  It reboots if it works (it terminates the navi software to update it, so the reboot is needed to restart it all up again).  It uses the USB root as a working area, so there will be files left there once it's done.  New and old versions of the exe and various manifests it updates.

Run this package with the Recore button in FG2ICCComms.exe